Refactored plugin logic structure a bit
