+++ packages/altair-core/src/index.ts
@@ -1,5 +1,5 @@
-import { PluginClass } from './plugin/plugin.interfaces';
+import { PluginBase, PluginConstructor } from './plugin/base';
 
-export const registerPluginClass = (pluginClassName: string, pluginClass: PluginClass) => {
+export const registerPluginClass = <T extends PluginBase>(pluginClassName: string, pluginClass: PluginConstructor<T>) => {
   (window as any)['AltairGraphQL'].plugins[pluginClassName] = pluginClass
 };
